requirements.txt
numpy
python-dotenv
talib